requirements.txt
pandas
pymysql 
SQLAlchemy 
sqlalchemy_utils
pymongo
psycopg2-binary